requirements.txt
streamlit==1.32.2
tensorflow>=2.15.0
numpy
opencv-python-headless
Pillow
matplotlib
openai
plotly
scikit-learn
